+++ aiohttp_things/views.py
@@ -1,4 +1,3 @@
-import warnings
 from abc import ABCMeta
 from typing import Any, Callable, Dict, Iterable
 
@@ -24,6 +23,18 @@ class ContextMixin(AbstractView, metaclass=ABCMeta):
         self.context = {}
 
 
+class PaginationMixin(AbstractView, metaclass=ABCMeta):
+    page: Any = None
+    page_adapter: Callable[..., Any]
+
+    def __init__(self, request: Request) -> None:
+        super().__init__(request)
+        page = self.request.rel_url.query.get('page')
+        if page:
+            page_adapter = getattr(self, 'page_adapter', lambda v: v)
+            self.page = page_adapter(page)
+
+
 class PrimaryKeyMixin(AbstractView, metaclass=ABCMeta):
     """
     Class based view mixin with primary key attribute.
@@ -34,30 +45,20 @@ class PrimaryKeyMixin(AbstractView, metaclass=ABCMeta):
 
     #: Callable object for converting a primary key.
     pk_adapter: Callable[..., Any]
-    pk_factory: Callable[..., Any]
 
     def __init__(self, request: Request) -> None:
         super().__init__(request)
         pk = self.request.match_info.get('pk')
-        if hasattr(self, 'pk_factory'):
-            msg = (
-                "The `PrimaryKeyMixin.pk_factory` attribute has been "
-                "deprecated. Use the `PrimaryKeyMixin.pk_adapter` attribute."
-            )
-            warnings.warn(msg, DeprecationWarning, stacklevel=2)
         if pk:
-            pk_factory = getattr(self, 'pk_factory', lambda v: v)
-            pk_adapter = getattr(self, 'pk_adapter', pk_factory)
+            pk_adapter = getattr(self, 'pk_adapter', lambda v: v)
             self.pk = pk_adapter(pk)
 
 
 class ItemMixin(AbstractView, metaclass=ABCMeta):
-    instance: Any
     item: Any
 
     def __init__(self, request: Request) -> None:
         super().__init__(request)
-        self.instance = None
         self.item = None