+++ example.py
@@ -1,7 +1,7 @@
 import json
 import uuid
+import aiohttp_things as ahth
 from aiohttp import web
-from aiohttp_things.views import JSONMixin, PrimaryKeyMixin
 
 
 def safe_json_value(value):
@@ -12,7 +12,7 @@ def safe_json_value(value):
         return str(value)
 
 
-class Base(web.View, JSONMixin, PrimaryKeyMixin):
+class Base(web.View, ahth.JSONMixin, ahth.PrimaryKeyMixin):
     async def get(self):
         self.context['Type of primary key'] = safe_json_value(type(self.pk))
         self.context['Value of primary key'] = safe_json_value(self.pk)